pyproject.toml
[project]
name = "adaptive-agent"
version = "0.1.0"
description = "Agent benchmarking framework for self-improvement strategies"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.39.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
